updates : - returns back Android Launcher - add grid_view_sapi.xml - update role privilege - add util moneyFormatter - add function to openWhatsApp on BaseFragment.kt

+++ app/src/main/java/com/example/sapiii/feature/detail/view/EditDataHewanActivity.kt
@@ -151,7 +151,7 @@ class EditDataHewanActivity : BaseActivity() {
             ),
             data = dataKambing.data.copy(
                 usia = binding.etUsiaskg.text.toString(),
-                beratBadan = binding.etBrtbdnSkg.text.toString().toInt()
+                beratBadan = binding.etBrtbdnSkg.text.toString().toDouble()
             )
         )
 
@@ -181,7 +181,7 @@ class EditDataHewanActivity : BaseActivity() {
             ),
             data = dataSapi.data.copy(
                 usia = binding.etUsiaskg.text.toString(),
-                beratBadan = binding.etBrtbdnSkg.text.toString().toInt()
+                beratBadan = binding.etBrtbdnSkg.text.toString().toDouble()
             )
         )
